+++ NEWS.md
@@ -1,3 +1,20 @@
+# diseq 0.0.14
+
+* Added option and documentation for estimating clustered standards errors.
+
+# diseq 0.0.13.9002
+
+* Added documentation for the function that return the scores.
+
+# diseq 0.0.13.9001
+
+* Added option for estimating heteroscedasticity-consistent (Huber-White) standard errors. 
+* Added functionality for extracting the score matrices of the estimated models. 
+
+# diseq 0.0.13.9000
+
+* Corrected documentation typos.
+
 # diseq 0.0.13
 
 * Corrections of non-canonical web-links in README. Adjustments before CRAN submission.